Implemented ClipGUI using SerializedObject
