+++ Scripts/BlendShape/Editor/BlendShapeClipEditorHelper.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: fcb56a7eb1db73c4cb9ea1689635b246
+timeCreated: 1541144400
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
