+++ Scripts/BlendShape/Editor/SerializedBlendShapeClipEditor.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using UnityEditor;
 using UnityEditorInternal;
 using UnityEngine;
@@ -20,6 +21,8 @@ namespace VRM
         #region BlendShapeBind
         public static int BlendShapeBindingHeight = 60;
         ReorderableList m_ValuesList;
+
+        SerializedProperty m_valuesProp;
         #endregion
 
         #region  MaterialValueBind
@@ -35,8 +38,11 @@ namespace VRM
         int m_mode;
         static string[] MODES = new[]{
             "BlendShape",
-            "Material"
+            "BlendShape List",
+            "Material List"
         };
+
+        MeshPreviewItem[] m_items;
         #endregion
 
         public SerializedBlendShapeEditor(SerializedObject serializedObject,
@@ -57,17 +63,17 @@ namespace VRM
 
             m_BlendShapeNameProp = serializedObject.FindProperty("BlendShapeName");
             m_PresetProp = serializedObject.FindProperty("Preset");
-            var valuesProp = serializedObject.FindProperty("Values");
+            m_valuesProp = serializedObject.FindProperty("Values");
 
-            m_ValuesList = new ReorderableList(serializedObject, valuesProp);
+            m_ValuesList = new ReorderableList(serializedObject, m_valuesProp);
             m_ValuesList.elementHeight = BlendShapeBindingHeight;
             m_ValuesList.drawElementCallback =
               (rect, index, isActive, isFocused) =>
               {
-                  var element = valuesProp.GetArrayElementAtIndex(index);
+                  var element = m_valuesProp.GetArrayElementAtIndex(index);
                   rect.height -= 4;
                   rect.y += 2;
-                  if (DrawBlendShapeBinding(rect, element, previewSceneManager))
+                  if (BlendShapeClipEditorHelper.DrawBlendShapeBinding(rect, element, previewSceneManager))
                   {
                       m_changed = true;
                   }
@@ -82,11 +88,15 @@ namespace VRM
                   var element = materialValuesProp.GetArrayElementAtIndex(index);
                   rect.height -= 4;
                   rect.y += 2;
-                  if (DrawMaterialValueBinding(rect, element, previewSceneManager))
+                  if (BlendShapeClipEditorHelper.DrawMaterialValueBinding(rect, element, previewSceneManager))
                   {
                       m_changed = true;
                   }
               };
+
+            m_items = previewSceneManager.EnumRenderItems
+            .Where(x => x.SkinnedMeshRenderer != null)
+            .ToArray();
         }
 
         public struct DrawResult
@@ -115,6 +125,13 @@ namespace VRM
             m_serializedObject.Update();
 
             EditorGUILayout.Space();
+
+            // ReadonlyのBlendShapeClip参照
+            GUI.enabled = false;
+            EditorGUILayout.ObjectField("Current clip",
+                m_targetObject, typeof(BlendShapeClip), false);
+            GUI.enabled = true;
+
             EditorGUILayout.PropertyField(m_BlendShapeNameProp, true);
             EditorGUILayout.PropertyField(m_PresetProp, true);
 
@@ -125,12 +142,18 @@ namespace VRM
             {
                 case 0:
                     {
+                        ClipGUI();
+                    }
+                    break;
+
+                case 1:
+                    {
                         //EditorGUILayout.LabelField("BlendShapeBindings", EditorStyles.boldLabel);
                         m_ValuesList.DoLayoutList();
                     }
                     break;
 
-                case 1:
+                case 2:
                     {
                         //EditorGUILayout.LabelField("MaterialValueBindings", EditorStyles.boldLabel);
                         m_MaterialValuesList.DoLayoutList();
@@ -152,38 +175,11 @@ namespace VRM
             };
         }
 
-#if false
-        void ClipGUI(BlendShapeClip clip)
+        void ClipGUI()
         {
-            if (clip == null)
-            {
-                Debug.LogWarning("no clip");
-                return;
-            }
-
-            EditorGUILayout.Space();
-            EditorGUILayout.LabelField("CurrentClip", EditorStyles.boldLabel);
-
-            // ReadonlyのBlendShapeClip参照
-            GUI.enabled = false;
-            EditorGUILayout.ObjectField("Current clip",
-                clip, typeof(BlendShapeClip), false);
-            GUI.enabled = true;
-
-            // Preset選択
-            clip.Preset = (BlendShapePreset)EditorGUILayout.Popup("Preset", (int)clip.Preset, Presets);
-
-            // Readonlyの名前入力
-            GUI.enabled = false;
-            EditorGUILayout.TextField("BlendShapeName", clip.BlendShapeName);
-            GUI.enabled = true;
-
             // Key重複の警告
-            m_selector.DuplicateWarn();
 
-            EditorGUILayout.Space();
-            EditorGUILayout.LabelField("BlendShapeValues", EditorStyles.boldLabel);
-
+            /*
             EditorGUILayout.BeginHorizontal();
             if (GUILayout.Button("Clear"))
             {
@@ -197,25 +193,60 @@ namespace VRM
                 EditorUtility.SetDirty(clip);
             }
             EditorGUILayout.EndHorizontal();
+            */
 
-            if (PreviewSceneManager != null)
+            var changed = BlendShapeBindsGUI();
+            if (changed)
             {
-                if (BlendShapeBindsGUI(clip))
+                string maxWeightName;
+                var bindings = GetBindings(out maxWeightName);
+                m_valuesProp.ClearArray();
+                m_valuesProp.arraySize = bindings.Length;
+                for (int i = 0; i < bindings.Length; ++i)
                 {
-                    PreviewSceneManager.Bake();
+                    var item = m_valuesProp.GetArrayElementAtIndex(i);
+
+                    var endProperty = item.GetEndProperty();
+                    while (item.NextVisible(true))
+                    {
+                        if (SerializedProperty.EqualContents(item, endProperty))
+                        {
+                            break;
+                        }
+
+                        switch (item.name)
+                        {
+                            case "RelativePath":
+                                item.stringValue = bindings[i].RelativePath;
+                                break;
+
+                            case "Index":
+                                item.intValue = bindings[i].Index;
+                                break;
+
+                            case "Weight":
+                                item.floatValue = bindings[i].Weight;
+                                break;
+
+                            default:
+                                throw new Exception();
+                        }
+                    }
                 }
+
+                m_changed = true;
             }
         }
 
         List<bool> m_meshFolds = new List<bool>();
-        bool BlendShapeBindsGUI(BlendShapeClip clip)
+        bool BlendShapeBindsGUI()
         {
             bool changed = false;
             int foldIndex = 0;
             // すべてのSkinnedMeshRendererを列挙する
-            foreach (var item in PreviewSceneManager.EnumRenderItems.Where(x => x.SkinnedMeshRenderer != null))
+            foreach (var renderer in m_items.Select(x => x.SkinnedMeshRenderer))
             {
-                var mesh = item.SkinnedMeshRenderer.sharedMesh;
+                var mesh = renderer.sharedMesh;
                 if (mesh != null && mesh.blendShapeCount > 0)
                 {
                     //var relativePath = UniGLTF.UnityExtensions.RelativePathFrom(renderer.transform, m_target.transform);
@@ -225,17 +256,17 @@ namespace VRM
                     {
                         m_meshFolds.Add(false);
                     }
-                    m_meshFolds[foldIndex] = EditorGUILayout.Foldout(m_meshFolds[foldIndex], item.SkinnedMeshRenderer.name);
+                    m_meshFolds[foldIndex] = EditorGUILayout.Foldout(m_meshFolds[foldIndex], renderer.name);
                     if (m_meshFolds[foldIndex])
                     {
                         //EditorGUI.indentLevel += 1;
                         for (int i = 0; i < mesh.blendShapeCount; ++i)
                         {
-                            var src = item.SkinnedMeshRenderer.GetBlendShapeWeight(i);
+                            var src = renderer.GetBlendShapeWeight(i);
                             var dst = EditorGUILayout.Slider(mesh.GetBlendShapeName(i), src, 0, 100.0f);
                             if (dst != src)
                             {
-                                item.SkinnedMeshRenderer.SetBlendShapeWeight(i, dst);
+                                renderer.SetBlendShapeWeight(i, dst);
                                 changed = true;
                             }
                         }
@@ -246,219 +277,48 @@ namespace VRM
             }
             return changed;
         }
-        private void ClearBlendShape()
+
+        BlendShapeBinding[] GetBindings(out string _maxWeightName)
         {
-            foreach (var item in PreviewSceneManager.EnumRenderItems.Where(x => x.SkinnedMeshRenderer != null))
+            var maxWeight = 0.0f;
+            var maxWeightName = "";
+            // weightのついたblendShapeを集める
+            var values = m_items
+                .SelectMany(x =>
             {
-                var renderer = item.SkinnedMeshRenderer;
-                var mesh = renderer.sharedMesh;
+                var mesh = x.SkinnedMeshRenderer.sharedMesh;
+
+                var relativePath = x.Path;
+
+                var list = new List<BlendShapeBinding>();
                 if (mesh != null)
                 {
                     for (int i = 0; i < mesh.blendShapeCount; ++i)
                     {
-                        renderer.SetBlendShapeWeight(i, 0);
+                        var weight = x.SkinnedMeshRenderer.GetBlendShapeWeight(i);
+                        if (weight == 0)
+                        {
+                            continue;
+                        }
+                        var name = mesh.GetBlendShapeName(i);
+                        if (weight > maxWeight)
+                        {
+                            maxWeightName = name;
+                            maxWeight = weight;
+                        }
+                        list.Add(new BlendShapeBinding
+                        {
+                            Index = i,
+                            RelativePath = relativePath,
+                            Weight = weight
+                        });
                     }
                 }
-            }
+                return list;
+            }).ToArray()
+            ;
+            _maxWeightName = maxWeightName;
+            return values;
         }
-#endif
-
-        #region private
-        static bool DrawBlendShapeBinding(Rect position, SerializedProperty property,
-            PreviewSceneManager scene)
-        {
-            bool changed = false;
-            if (scene != null)
-            {
-                var height = 16;
-
-                var y = position.y;
-                var rect = new Rect(position.x, y, position.width, height);
-                int pathIndex;
-                if (StringPopup(rect, property.FindPropertyRelative("RelativePath"), scene.SkinnedMeshRendererPathList, out pathIndex))
-                {
-                    changed = true;
-                }
-
-                y += height;
-                rect = new Rect(position.x, y, position.width, height);
-                int blendShapeIndex;
-                if (IntPopup(rect, property.FindPropertyRelative("Index"), scene.GetBlendShapeNames(pathIndex), out blendShapeIndex))
-                {
-                    changed = true;
-                }
-
-                y += height;
-                rect = new Rect(position.x, y, position.width, height);
-                if (FloatSlider(rect, property.FindPropertyRelative("Weight"), 100))
-                {
-                    changed = true;
-                }
-            }
-            return changed;
-        }
-
-        static bool DrawMaterialValueBinding(Rect position, SerializedProperty property,
-            PreviewSceneManager scene)
-        {
-            bool changed = false;
-            if (scene != null)
-            {
-                var height = 16;
-
-                var y = position.y;
-                var rect = new Rect(position.x, y, position.width, height);
-                int materialIndex;
-                if (StringPopup(rect, property.FindPropertyRelative("MaterialName"), scene.MaterialNames, out materialIndex))
-                {
-                    changed = true;
-                }
-
-                if (materialIndex >= 0)
-                {
-                    var materialItem = scene.GetMaterialItem(scene.MaterialNames[materialIndex]);
-                    if (materialItem != null)
-                    {
-                        y += height;
-                        rect = new Rect(position.x, y, position.width, height);
-
-                        // プロパティ名のポップアップ
-                        int propIndex;
-                        if (StringPopup(rect, property.FindPropertyRelative("ValueName"), materialItem.PropNames, out propIndex))
-                        {
-                            changed = true;
-                        }
-
-                        if (propIndex >= 0)
-                        {
-                            // 有効なプロパティ名が選択された
-                            var propItem = materialItem.PropMap[materialItem.PropNames[propIndex]];
-                            {
-                                switch (propItem.PropertyType)
-                                {
-                                    case ShaderUtil.ShaderPropertyType.Color:
-                                        {
-                                            property.FindPropertyRelative("BaseValue").vector4Value = propItem.DefaultValues;
-
-                                            // max
-                                            y += height;
-                                            rect = new Rect(position.x, y, position.width, height);
-                                            if (ColorProp(rect, property.FindPropertyRelative("TargetValue")))
-                                            {
-                                                changed = true;
-                                            }
-                                        }
-                                        break;
-
-                                    case ShaderUtil.ShaderPropertyType.TexEnv:
-                                        {
-                                            property.FindPropertyRelative("BaseValue").vector4Value = propItem.DefaultValues;
-
-                                            // max
-                                            y += height;
-                                            rect = new Rect(position.x, y, position.width, height);
-                                            if (OffsetProp(rect, property.FindPropertyRelative("TargetValue")))
-                                            {
-                                                changed = true;
-                                            }
-                                        }
-                                        break;
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            return changed;
-        }
-
-
-        static bool StringPopup(Rect rect, SerializedProperty prop, string[] options, out int newIndex)
-        {
-            if (options == null)
-            {
-                newIndex = -1;
-                return false;
-            }
-
-            var oldIndex = Array.IndexOf(options, prop.stringValue);
-            newIndex = EditorGUI.Popup(rect, oldIndex, options);
-            if (newIndex != oldIndex && newIndex >= 0 && newIndex < options.Length)
-            {
-                prop.stringValue = options[newIndex];
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        static bool IntPopup(Rect rect, SerializedProperty prop, string[] options, out int newIndex)
-        {
-            if (options == null)
-            {
-                newIndex = -1;
-                return false;
-            }
-
-            var oldIndex = prop.intValue;
-            newIndex = EditorGUI.Popup(rect, oldIndex, options);
-            if (newIndex != oldIndex && newIndex >= 0 && newIndex < options.Length)
-            {
-                prop.intValue = newIndex;
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        static bool FloatSlider(Rect rect, SerializedProperty prop, float maxValue)
-        {
-            var oldValue = prop.floatValue;
-            var newValue = EditorGUI.Slider(rect, prop.floatValue, 0, 100f);
-            if (newValue != oldValue)
-            {
-                prop.floatValue = newValue;
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        static bool ColorProp(Rect rect, SerializedProperty prop)
-        {
-            var oldValue = (Color)prop.vector4Value;
-            var newValue = EditorGUI.ColorField(rect, prop.displayName, oldValue);
-            if (newValue != oldValue)
-            {
-                prop.vector4Value = newValue;
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        static bool OffsetProp(Rect rect, SerializedProperty prop)
-        {
-            var oldValue = prop.vector4Value;
-            var newValue = EditorGUI.Vector4Field(rect, prop.displayName, oldValue);
-            if (newValue != oldValue)
-            {
-                prop.vector4Value = newValue;
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-        #endregion
     }
 }
